Allow builds if one of the backends is disabled
After the backend rework the build required that all backends are enabled
in every case. If one of the headers was disabled a missing header error
would be reported.

By moving the inclusion of the headers to the gepard-engine.ccp file and
guarding them correctly the build problem can be resolved.

Signed-off-by: Peter Gal <galpeter@inf.u-szeged.hu>